Create Cat Toy
This project is a prototype of a cat toy, which uses an ultrasonic senser to track the distance of a cat. When it is within a certain range, the servo begins spinning, which allows the cat to play with the toy, being a string which is attached to the servo.